ADDED: The code is normal and premium rides generation

+++ src/main/java/InvoiceSummary.java
@@ -1,9 +1,9 @@
 public class InvoiceSummary {
     double avgFare;
-    int totalFare;
+    double totalFare;
     double numOfRides;
 
-    public InvoiceSummary(double numOfRides, int totalFare) {
+    public InvoiceSummary(double numOfRides,  double totalFare) {
         this.numOfRides = numOfRides;
         this.totalFare = totalFare;
         this.avgFare = this.numOfRides / this.totalFare;
